Error update part 4: Rename the remaining types for consistency
* InvalidUtf16FirstUnit -> Utf16FirstUnitError
* InvalidUtf16Array -> Utf16ArrayError
* InvalidUtf16Slice -> Utf16SliceError
* InvalidUtf16Tuple -> Utf16TupleError

Mostly search & replace.

The naming is still a bit inconsistent in that Utf*Error has the
opposite meaning of NonAsciiError, NonBmpError and EmptyStrError.
But those are empty structs with no variant to tell what's wrong,
so it's not entirely arbitrary.

I also considered renaming CodepointError back to InvalidCodepont
to get consistency of the enums that way,
but there's already Utf16PairError so I decided to complete
what I started when I created that one.